requirements-test.txt
# Dependências para testes
pytest==7.4.0
pytest-django==4.5.2
pytest-mock==3.11.1
factory-boy==3.2.1
Pillow==10.0.0
fpdf2==2.7.4
PyMuPDF==1.22.5
rarfile==4.0
py7zr==0.20.5
